refactor(firebase): move core functionality to base project
- define get_project_specifics_for_firebase for find project
- use old_id instead of id when syncing models with firebase
- handle specific exception when handling invalid project config
- move common functions to project_types/firebase
